+++ src/main/java/me/khajiitos/jackseconomy/JacksEconomyClient.java
@@ -1,6 +1,7 @@
 package me.khajiitos.jackseconomy;
 
 import com.mojang.blaze3d.platform.InputConstants;
+import com.simibubi.create.foundation.ponder.CreatePonderPlugin;
 import me.khajiitos.jackseconomy.create.CreateCheck;
 import me.khajiitos.jackseconomy.create.CreateClient;
 import me.khajiitos.jackseconomy.create.CreatePonder;
@@ -13,6 +14,7 @@ import me.khajiitos.jackseconomy.listener.ClientRenderEventListeners;
 import me.khajiitos.jackseconomy.data.price.ItemDescription;
 import me.khajiitos.jackseconomy.data.price.PricesItemPriceInfo;
 import me.khajiitos.jackseconomy.screen.*;
+import net.createmod.ponder.foundation.PonderIndex;
 import net.minecraft.client.KeyMapping;
 import net.minecraft.client.gui.screens.MenuScreens;
 import net.minecraftforge.client.ConfigScreenHandler;
@@ -45,7 +47,7 @@ public class JacksEconomyClient {
 
         ModLoadingContext.get().registerExtensionPoint(ConfigScreenHandler.ConfigScreenFactory.class, () -> new ConfigScreenHandler.ConfigScreenFactory(ClientConfigScreen::new));
 
-        if (CreateCheck.isInstalled()) { CreatePonder.register(); }
+        if (CreateCheck.isInstalled()) { PonderIndex.addPlugin(new CreatePonder()); }
     }
 
     public static void onClientSetup(FMLClientSetupEvent e) {